pyproject.toml
[project]
name = "photo-organiser"
version = "0.1.0"
description = "Organise and make sense of all your photos."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "alembic>=1.12.0",
    "pydantic>=2.4.0",
    "deepface>=0.0.79",
    "easyocr>=1.7.0",
    "chromadb>=0.4.15",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ftfy>=6.1.0",
    "regex>=2023.8.8",
    "psutil>=5.9.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[mypy]>=2.0.45",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "black>=23.9.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
photo-organiser = "photo_organiser.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "ISC", "ICN", "G", "PIE", "PYI", "RUF", "SIM"]
ignore = ["E501", "N805"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.11",
]
